refactor: separate server/client logic in MyPage for hybrid SSR-CSR rendering
